refact: Remove ActivitySkeleton and add ActivityItemSkeleton
