+++ lib/components/profile/activity/index.ts
@@ -1,4 +1,3 @@
 import Activity from "./Activity";
-import ActivitySkeleton from "./ActivitySkeleton";
 
-export { Activity, ActivitySkeleton};
+export { Activity };
