+++ lib/components/profile/index.ts
@@ -1,6 +1,6 @@
-import { Activity, ActivitySkeleton } from "./activity";
+import { Activity } from "./activity";
 import { Badges } from "./badge";
 import PointsView from "./PointsView";
 import ProfileHeader from "./ProfileHeader";
 
-export { Activity, ActivitySkeleton, Badges, PointsView, ProfileHeader };
+export { Activity, Badges, PointsView, ProfileHeader };
